+++ assets/js/Components/App.js
@@ -11,7 +11,7 @@ import SettingsPage from './SettingsPage'
 import Api from '../Services/Api'
 import MessageTray from './Widgets/MessageTray'
 import { analyzeReport } from '../Services/Report'
-import { DEFAULT_USER_SETTINGS, UFIXIT_OPTIONS } from '../Services/Settings'
+import { DEFAULT_USER_SETTINGS } from '../Services/Settings'
 import { ISSUE_STATE } from '../Services/Constants'
 
 
@@ -20,10 +20,7 @@ export default function App(initialData) {
   const [nextMessage, setNextMessage] = useState('')
   const [untranslatedMessage, setUntranslatedMessage] = useState('')
   const [report, setReport] = useState(initialData.report || null)  
-  const [settings, setSettings] = useState(Object.assign({},
-    initialData?.settings || {},
-    { UFIXIT_OPTIONS },
-  ))
+  const [settings, setSettings] = useState({...initialData.settings});
   const [preferences, setPreferences] = useState(initialData.preferences || null);
   const [textSpacing, setTextSpacing] = useState(preferences.textSpacing ?? DEFAULT_USER_SETTINGS.TEXT_SPACING) 
   const [sections, setSections] = useState([])
