+++ assets/js/Components/Forms/AltText.js
@@ -3,6 +3,7 @@ import RadioSelector from '../Widgets/RadioSelector'
 import OptionFeedback from '../Widgets/OptionFeedback'
 import * as Html from '../../Services/Html'
 import * as Text from '../../Services/Text'
+import { UFIXIT_OPTIONS } from '../../Services/Constants'
 
 export default function AltText ({
   t,
@@ -17,9 +18,9 @@ export default function AltText ({
  }) {
   
   const FORM_OPTIONS = {
-    ADD_TEXT: settings.UFIXIT_OPTIONS.ADD_TEXT,
-    MARK_DECORATIVE: settings.UFIXIT_OPTIONS.MARK_DECORATIVE,
-    MARK_AS_REVIEWED: settings.UFIXIT_OPTIONS.MARK_AS_REVIEWED
+    ADD_TEXT: UFIXIT_OPTIONS.ADD_TEXT,
+    MARK_DECORATIVE: UFIXIT_OPTIONS.MARK_DECORATIVE,
+    MARK_AS_REVIEWED: UFIXIT_OPTIONS.MARK_AS_REVIEWED
   }
   const maxLength = 150
