+++ assets/js/Components/Forms/HeadingStyleForm.js
@@ -4,6 +4,7 @@ import OptionFeedback from '../Widgets/OptionFeedback'
 import Combobox from '../Widgets/Combobox'
 import ToggleSwitch from '../Widgets/ToggleSwitch'
 import * as Html from '../../Services/Html'
+import { UFIXIT_OPTIONS } from '../../Services/Constants'
 
 export default function HeadingStyleForm ({
   t,
@@ -21,8 +22,8 @@ export default function HeadingStyleForm ({
   const tagOptions = ["H2", "H3", "H4", "H5", "H6"]
   const allHeadings = ["H1", "H2", "H3", "H4", "H5", "H6", "h1", "h2", "h3", "h4", "h5", "h6"]
   const FORM_OPTIONS = {
-    SELECT_LEVEL: settings.UFIXIT_OPTIONS.SELECT_TAG,
-    MARK_AS_REVIEWED: settings.UFIXIT_OPTIONS.MARK_AS_REVIEWED
+    SELECT_LEVEL: UFIXIT_OPTIONS.SELECT_TAG,
+    MARK_AS_REVIEWED: UFIXIT_OPTIONS.MARK_AS_REVIEWED
   }
   
   const [selectOptions, setSelectOptions] = useState([])
